Google sign in with arctic and lucia

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -1,6 +1,7 @@
 import { validateRequest } from "@/auth";
+import { notifyUserCreated , notifyUserDeleted  } from "@/lib/notifySocket";
 import prisma from "@/lib/prisma";
-import { LikeInfo } from "@/lib/types";
+import { LikeInfo, notificationsInclude } from "@/lib/types";
 
 export async function GET(
   req: Request,
@@ -60,7 +61,19 @@ export async function POST(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.upsert({
+    // the user has liked certain post.
+
+    const post = await prisma.post.findUnique({
+      where: {
+        id: postId,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "User not found" }, { status: 404 });
+    }
+
+    const upsertLike = prisma.like.upsert({
       where: {
         postId_userId: {
           postId: postId,
@@ -73,6 +86,25 @@ export async function POST(
       },
       update: {},
     });
+
+    const createNotification = prisma.notification.create({
+      data: {
+        issuerId: loggedInUser.id,
+        recipientId: post.userId,
+        postId: post.id,
+        type: "LIKE",
+      },
+      include: notificationsInclude,
+    });
+
+    const [_newLike, notification] = await prisma.$transaction([
+      upsertLike,
+      ...(post.userId !== loggedInUser.id ? [createNotification] : []),
+    ]);
+
+    if (post.userId !== loggedInUser.id) {
+      await notifyUserCreated (post.userId, notification);
+    }
 
     return new Response();
   } catch (error) {
@@ -92,12 +124,50 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.deleteMany({
+    const post = await prisma.post.findUnique({
+      where: {
+        id: postId,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "User not found" }, { status: 404 });
+    }
+
+    const deleteLike = prisma.like.deleteMany({
       where: {
         postId: postId,
         userId: loggedInUser.id,
       },
     });
+
+    // we dont have the id of the like to be deleted. Delete wont work without it. therefore using deleteMany
+    const findNotification = prisma.notification.findFirst({
+      where: {
+        issuerId: loggedInUser.id,
+        recipientId: post.userId,
+        type: "LIKE",
+        postId: post.id,
+      },
+    });
+
+    // deleteMany need not be in a condition because if there is no notification if will just be ignored.
+    // notification is not created when the user likes his own post, therefore there is nothing to delete. hence no condition needed.
+
+    const [_deletedLike, linkedNnotification] = await prisma.$transaction([
+      deleteLike,
+      findNotification,
+    ]);
+
+    const deletedNotification = await prisma.notification.delete({
+      where: {
+        id: linkedNnotification?.id,
+      },
+    });
+
+    if (post.userId !== loggedInUser.id) {
+      await notifyUserDeleted (post.userId, deletedNotification.id);
+    }
 
     return new Response();
   } catch (error) {
